+++ docs/CHANGELOG.rst
@@ -13,6 +13,7 @@ Changed
 -------
 - Add trimming with Trimmomatic in ``workflow-wgbs-single`` and
   ``workflow-wgbs-paired`` workflows
+- Make intervals an optional input in ``bqsr`` process
 
 Fixed
 -------